+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## `v0.0.71`
+
+### Features
+
+- Add optional `label` field to `ConnectedWallet` instances, which contains the user-defined label of the wallet
+
+### Fixes
+
+- Handle more wallet errors in the `wrap` function
+
 ## `v0.0.70`
 
 ### Features